+++ crates/swc_ecma_parser/src/lexer/token_table.rs
